+++ lib/src/compatible/compatible_smart_dialog.dart
@@ -126,6 +126,7 @@ class CompatibleSmartDialog {
       animationType: isLoadingTemp ?? config.isLoading
           ? SmartAnimationType.centerFade_otherSlide
           : SmartAnimationType.centerScale_otherSlide,
+      nonAnimationTypes: const [],
       animationBuilder: null,
       usePenetrate: isPenetrateTemp ?? config.isPenetrate,
       useAnimation: isUseAnimationTemp ?? config.isUseAnimation,
@@ -283,6 +284,7 @@ class CompatibleSmartDialog {
       animationType: isLoadingTemp ?? false
           ? SmartAnimationType.fade
           : SmartAnimationType.scale,
+      nonAnimationTypes: const [],
       animationBuilder: null,
       scalePointBuilder: null,
       usePenetrate: isPenetrateTemp ?? config.isPenetrate,
@@ -381,6 +383,7 @@ class CompatibleSmartDialog {
       animationType: isLoadingTemp ?? true
           ? SmartAnimationType.fade
           : SmartAnimationType.scale,
+      nonAnimationTypes: const [],
       animationBuilder: null,
       usePenetrate: isPenetrateTemp ?? false,
       useAnimation: isUseAnimationTemp ?? config.isUseAnimation,